bpffeature: set version for older kernels
